+++ packages/go_router_builder/test/test_inputs/_go_router_builder_test_input.dart
@@ -54,15 +54,6 @@ class NullableRequiredParam extends GoRouteData {
 }
 
 @ShouldThrow(
-  r'Parameters named `$extra` cannot be required.',
-)
-@TypedGoRoute<ExtraMustBeOptional>(path: r'bob/:$extra')
-class ExtraMustBeOptional extends GoRouteData {
-  ExtraMustBeOptional({required this.$extra});
-  final int $extra;
-}
-
-@ShouldThrow(
   'Missing param `id` in path.',
 )
 @TypedGoRoute<MissingPathParam>(path: 'bob/')
@@ -72,7 +63,7 @@ class MissingPathParam extends GoRouteData {
 }
 
 @ShouldGenerate(r'''
-GoRoute get $enumParam => GoRouteData.$route(
+RouteBase get $enumParam => GoRouteData.$route(
       path: '/:y',
       factory: $EnumParamExtension._fromState,
     );
@@ -88,7 +79,7 @@ extension $EnumParamExtension on EnumParam {
 
   void go(BuildContext context) => context.go(location);
 
-  void push(BuildContext context) => context.push(location);
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
 
   void pushReplacement(BuildContext context) =>
       context.pushReplacement(location);
@@ -121,7 +112,7 @@ enum EnumTest {
 }
 
 @ShouldGenerate(r'''
-GoRoute get $defaultValueRoute => GoRouteData.$route(
+RouteBase get $defaultValueRoute => GoRouteData.$route(
       path: '/default-value-route',
       factory: $DefaultValueRouteExtension._fromState,
     );
@@ -140,7 +131,7 @@ extension $DefaultValueRouteExtension on DefaultValueRoute {
 
   void go(BuildContext context) => context.go(location);
 
-  void push(BuildContext context) => context.push(location);
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
 
   void pushReplacement(BuildContext context) =>
       context.pushReplacement(location);
@@ -162,7 +153,7 @@ class DefaultValueRoute extends GoRouteData {
 }
 
 @ShouldGenerate(r'''
-GoRoute get $extraValueRoute => GoRouteData.$route(
+RouteBase get $extraValueRoute => GoRouteData.$route(
       path: '/default-value-route',
       factory: $ExtraValueRouteExtension._fromState,
     );
@@ -182,7 +173,8 @@ extension $ExtraValueRouteExtension on ExtraValueRoute {
 
   void go(BuildContext context) => context.go(location, extra: $extra);
 
-  void push(BuildContext context) => context.push(location, extra: $extra);
+  Future<T?> push<T>(BuildContext context) =>
+      context.push<T>(location, extra: $extra);
 
   void pushReplacement(BuildContext context) =>
       context.pushReplacement(location, extra: $extra);
@@ -204,6 +196,37 @@ class ExtraValueRoute extends GoRouteData {
   final int? $extra;
 }
 
+@ShouldGenerate(r'''
+RouteBase get $requiredExtraValueRoute => GoRouteData.$route(
+      path: '/default-value-route',
+      factory: $RequiredExtraValueRouteExtension._fromState,
+    );
+
+extension $RequiredExtraValueRouteExtension on RequiredExtraValueRoute {
+  static RequiredExtraValueRoute _fromState(GoRouterState state) =>
+      RequiredExtraValueRoute(
+        $extra: state.extra as int,
+      );
+
+  String get location => GoRouteData.$location(
+        '/default-value-route',
+      );
+
+  void go(BuildContext context) => context.go(location, extra: $extra);
+
+  Future<T?> push<T>(BuildContext context) =>
+      context.push<T>(location, extra: $extra);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location, extra: $extra);
+}
+''')
+@TypedGoRoute<RequiredExtraValueRoute>(path: '/default-value-route')
+class RequiredExtraValueRoute extends GoRouteData {
+  RequiredExtraValueRoute({required this.$extra});
+  final int $extra;
+}
+
 @ShouldThrow(
   'Default value used with a nullable type. Only non-nullable type can have a default value.',
   todo: 'Remove the default value or make the type non-nullable.',
@@ -215,7 +238,7 @@ class NullableDefaultValueRoute extends GoRouteData {
 }
 
 @ShouldGenerate(r'''
-GoRoute get $iterableWithEnumRoute => GoRouteData.$route(
+RouteBase get $iterableWithEnumRoute => GoRouteData.$route(
       path: '/iterable-with-enum',
       factory: $IterableWithEnumRouteExtension._fromState,
     );
@@ -238,7 +261,7 @@ extension $IterableWithEnumRouteExtension on IterableWithEnumRoute {
 
   void go(BuildContext context) => context.go(location);
 
-  void push(BuildContext context) => context.push(location);
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
 
   void pushReplacement(BuildContext context) =>
       context.pushReplacement(location);
@@ -269,7 +292,7 @@ enum EnumOnlyUsedInIterable {
 }
 
 @ShouldGenerate(r'''
-GoRoute get $iterableDefaultValueRoute => GoRouteData.$route(
+RouteBase get $iterableDefaultValueRoute => GoRouteData.$route(
       path: '/iterable-default-value-route',
       factory: $IterableDefaultValueRouteExtension._fromState,
     );
@@ -291,7 +314,7 @@ extension $IterableDefaultValueRouteExtension on IterableDefaultValueRoute {
 
   void go(BuildContext context) => context.go(location);
 
-  void push(BuildContext context) => context.push(location);
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
 
   void pushReplacement(BuildContext context) =>
       context.pushReplacement(location);