+++ packages/local_auth/local_auth_android/android/src/main/java/io/flutter/plugins/localauth/LocalAuthPlugin.java
@@ -21,13 +21,18 @@ import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.embedding.engine.plugins.lifecycle.FlutterLifecycleAdapter;
-import io.flutter.plugin.common.MethodCall;
-import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
-import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry;
 import io.flutter.plugins.localauth.AuthenticationHelper.AuthCompletionHandler;
+import io.flutter.plugins.localauth.Messages.AuthClassification;
+import io.flutter.plugins.localauth.Messages.AuthClassificationWrapper;
+import io.flutter.plugins.localauth.Messages.AuthOptions;
+import io.flutter.plugins.localauth.Messages.AuthResult;
+import io.flutter.plugins.localauth.Messages.AuthResultWrapper;
+import io.flutter.plugins.localauth.Messages.AuthStrings;
+import io.flutter.plugins.localauth.Messages.LocalAuthApi;
+import io.flutter.plugins.localauth.Messages.Result;
 import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
@@ -35,9 +40,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
  *
  * <p>Instantiate this in an add to app scenario to gracefully handle activity and context changes.
  */
-@SuppressWarnings("deprecation")
-public class LocalAuthPlugin implements MethodCallHandler, FlutterPlugin, ActivityAware {
-  private static final String CHANNEL_NAME = "plugins.flutter.io/local_auth_android";
+public class LocalAuthPlugin implements FlutterPlugin, ActivityAware, LocalAuthApi {
   private static final int LOCK_REQUEST_CODE = 221;
   private Activity activity;
   private AuthenticationHelper authHelper;
@@ -45,20 +48,19 @@ public class LocalAuthPlugin implements MethodCallHandler, FlutterPlugin, Activi
   @VisibleForTesting final AtomicBoolean authInProgress = new AtomicBoolean(false);
 
   // These are null when not using v2 embedding.
-  private MethodChannel channel;
   private Lifecycle lifecycle;
   private BiometricManager biometricManager;
   private KeyguardManager keyguardManager;
-  private Result lockRequestResult;
+  Result<AuthResultWrapper> lockRequestResult;
   private final PluginRegistry.ActivityResultListener resultListener =
       new PluginRegistry.ActivityResultListener() {
         @Override
         public boolean onActivityResult(int requestCode, int resultCode, Intent data) {
           if (requestCode == LOCK_REQUEST_CODE) {
             if (resultCode == RESULT_OK && lockRequestResult != null) {
-              authenticateSuccess(lockRequestResult);
+              onAuthenticationCompleted(lockRequestResult, AuthResult.SUCCESS);
             } else {
-              authenticateFail(lockRequestResult);
+              onAuthenticationCompleted(lockRequestResult, AuthResult.FAILURE);
             }
             lockRequestResult = null;
           }
@@ -72,16 +74,13 @@ public class LocalAuthPlugin implements MethodCallHandler, FlutterPlugin, Activi
    * <p>Calling this will register the plugin with the passed registrar. However, plugins
    * initialized this way won't react to changes in activity or context.
    *
-   * @param registrar attaches this plugin's {@link
-   *     io.flutter.plugin.common.MethodChannel.MethodCallHandler} to the registrar's {@link
-   *     io.flutter.plugin.common.BinaryMessenger}.
+   * @param registrar provides access to necessary plugin context.
    */
   @SuppressWarnings("deprecation")
-  public static void registerWith(PluginRegistry.Registrar registrar) {
-    final MethodChannel channel = new MethodChannel(registrar.messenger(), CHANNEL_NAME);
+  public static void registerWith(@NonNull PluginRegistry.Registrar registrar) {
     final LocalAuthPlugin plugin = new LocalAuthPlugin();
     plugin.activity = registrar.activity();
-    channel.setMethodCallHandler(plugin);
+    LocalAuthApi.setup(registrar.messenger(), plugin);
     registrar.addActivityResultListener(plugin.resultListener);
   }
 
@@ -92,174 +91,114 @@ public class LocalAuthPlugin implements MethodCallHandler, FlutterPlugin, Activi
    */
   public LocalAuthPlugin() {}
 
-  @Override
-  public void onMethodCall(MethodCall call, @NonNull final Result result) {
-    switch (call.method) {
-      case "authenticate":
-        authenticate(call, result);
-        break;
-      case "getEnrolledBiometrics":
-        getEnrolledBiometrics(result);
-        break;
-      case "isDeviceSupported":
-        isDeviceSupported(result);
-        break;
-      case "stopAuthentication":
-        stopAuthentication(result);
-        break;
-      case "deviceSupportsBiometrics":
-        deviceSupportsBiometrics(result);
-        break;
-      default:
-        result.notImplemented();
-        break;
-    }
+  public @NonNull Boolean isDeviceSupported() {
+    return isDeviceSecure() || canAuthenticateWithBiometrics();
   }
 
-  /*
-   * Starts authentication process
-   */
-  private void authenticate(MethodCall call, final Result result) {
-    if (authInProgress.get()) {
-      result.error("auth_in_progress", "Authentication in progress", null);
-      return;
-    }
-
-    if (activity == null || activity.isFinishing()) {
-      result.error("no_activity", "local_auth plugin requires a foreground activity", null);
-      return;
-    }
-
-    if (!(activity instanceof FragmentActivity)) {
-      result.error(
-          "no_fragment_activity",
-          "local_auth plugin requires activity to be a FragmentActivity.",
-          null);
-      return;
-    }
-
-    if (!isDeviceSupported()) {
-      authInProgress.set(false);
-      result.error("NotAvailable", "Required security features not enabled", null);
-      return;
-    }
-
-    authInProgress.set(true);
-    AuthCompletionHandler completionHandler = createAuthCompletionHandler(result);
-
-    boolean isBiometricOnly = call.argument("biometricOnly");
-    boolean allowCredentials = !isBiometricOnly && canAuthenticateWithDeviceCredential();
-
-    sendAuthenticationRequest(call, completionHandler, allowCredentials);
-    return;
+  public @NonNull Boolean deviceCanSupportBiometrics() {
+    return hasBiometricHardware();
   }
 
-  @VisibleForTesting
-  public AuthCompletionHandler createAuthCompletionHandler(final Result result) {
-    return new AuthCompletionHandler() {
-      @Override
-      public void onSuccess() {
-        authenticateSuccess(result);
-      }
-
-      @Override
-      public void onFailure() {
-        authenticateFail(result);
-      }
-
-      @Override
-      public void onError(String code, String error) {
-        if (authInProgress.compareAndSet(true, false)) {
-          result.error(code, error, null);
-        }
-      }
-    };
-  }
-
-  @VisibleForTesting
-  public void sendAuthenticationRequest(
-      MethodCall call, AuthCompletionHandler completionHandler, boolean allowCredentials) {
-    authHelper =
-        new AuthenticationHelper(
-            lifecycle, (FragmentActivity) activity, call, completionHandler, allowCredentials);
-
-    authHelper.authenticate();
-  }
-
-  private void authenticateSuccess(Result result) {
-    if (authInProgress.compareAndSet(true, false)) {
-      result.success(true);
+  public @NonNull List<AuthClassificationWrapper> getEnrolledBiometrics() {
+    ArrayList<AuthClassificationWrapper> biometrics = new ArrayList<>();
+    if (biometricManager.canAuthenticate(BiometricManager.Authenticators.BIOMETRIC_WEAK)
+        == BiometricManager.BIOMETRIC_SUCCESS) {
+      biometrics.add(wrappedBiometric(AuthClassification.WEAK));
     }
-  }
-
-  private void authenticateFail(Result result) {
-    if (authInProgress.compareAndSet(true, false)) {
-      result.success(false);
+    if (biometricManager.canAuthenticate(BiometricManager.Authenticators.BIOMETRIC_STRONG)
+        == BiometricManager.BIOMETRIC_SUCCESS) {
+      biometrics.add(wrappedBiometric(AuthClassification.STRONG));
     }
+    return biometrics;
   }
 
-  /*
-   * Stops the authentication if in progress.
-   */
-  private void stopAuthentication(Result result) {
+  private @NonNull AuthClassificationWrapper wrappedBiometric(AuthClassification value) {
+    return new AuthClassificationWrapper.Builder().setValue(value).build();
+  }
+
+  public @NonNull Boolean stopAuthentication() {
     try {
       if (authHelper != null && authInProgress.get()) {
         authHelper.stopAuthentication();
         authHelper = null;
       }
       authInProgress.set(false);
-      result.success(true);
+      return true;
     } catch (Exception e) {
-      result.success(false);
+      return false;
     }
   }
 
-  private void deviceSupportsBiometrics(final Result result) {
-    result.success(hasBiometricHardware());
-  }
-
-  /*
-   * Returns enrolled biometric types available on device.
-   */
-  private void getEnrolledBiometrics(final Result result) {
-    try {
-      if (activity == null || activity.isFinishing()) {
-        result.error("no_activity", "local_auth plugin requires a foreground activity", null);
-        return;
-      }
-      ArrayList<String> biometrics = getEnrolledBiometrics();
-      result.success(biometrics);
-    } catch (Exception e) {
-      result.error("no_biometrics_available", e.getMessage(), null);
+  public void authenticate(
+      @NonNull AuthOptions options,
+      @NonNull AuthStrings strings,
+      @NonNull Result<AuthResultWrapper> result) {
+    if (authInProgress.get()) {
+      result.success(
+          new AuthResultWrapper.Builder().setValue(AuthResult.ERROR_ALREADY_IN_PROGRESS).build());
+      return;
     }
+
+    if (activity == null || activity.isFinishing()) {
+      result.success(
+          new AuthResultWrapper.Builder().setValue(AuthResult.ERROR_NO_ACTIVITY).build());
+      return;
+    }
+
+    if (!(activity instanceof FragmentActivity)) {
+      result.success(
+          new AuthResultWrapper.Builder().setValue(AuthResult.ERROR_NOT_FRAGMENT_ACTIVITY).build());
+      return;
+    }
+
+    if (!isDeviceSupported()) {
+      result.success(
+          new AuthResultWrapper.Builder().setValue(AuthResult.ERROR_NOT_AVAILABLE).build());
+      return;
+    }
+
+    authInProgress.set(true);
+    AuthCompletionHandler completionHandler = createAuthCompletionHandler(result);
+
+    boolean allowCredentials = !options.getBiometricOnly() && canAuthenticateWithDeviceCredential();
+
+    sendAuthenticationRequest(options, strings, allowCredentials, completionHandler);
   }
 
   @VisibleForTesting
-  public ArrayList<String> getEnrolledBiometrics() {
-    ArrayList<String> biometrics = new ArrayList<>();
-    if (activity == null || activity.isFinishing()) {
-      return biometrics;
+  public @NonNull AuthCompletionHandler createAuthCompletionHandler(
+      @NonNull final Result<AuthResultWrapper> result) {
+    return authResult -> onAuthenticationCompleted(result, authResult);
+  }
+
+  @VisibleForTesting
+  public void sendAuthenticationRequest(
+      @NonNull AuthOptions options,
+      @NonNull AuthStrings strings,
+      boolean allowCredentials,
+      @NonNull AuthCompletionHandler completionHandler) {
+    authHelper =
+        new AuthenticationHelper(
+            lifecycle,
+            (FragmentActivity) activity,
+            options,
+            strings,
+            completionHandler,
+            allowCredentials);
+
+    authHelper.authenticate();
+  }
+
+  void onAuthenticationCompleted(Result<AuthResultWrapper> result, AuthResult value) {
+    if (authInProgress.compareAndSet(true, false)) {
+      result.success(new AuthResultWrapper.Builder().setValue(value).build());
     }
-    if (biometricManager.canAuthenticate(BiometricManager.Authenticators.BIOMETRIC_WEAK)
-        == BiometricManager.BIOMETRIC_SUCCESS) {
-      biometrics.add("weak");
-    }
-    if (biometricManager.canAuthenticate(BiometricManager.Authenticators.BIOMETRIC_STRONG)
-        == BiometricManager.BIOMETRIC_SUCCESS) {
-      biometrics.add("strong");
-    }
-    return biometrics;
   }
 
   @VisibleForTesting
   public boolean isDeviceSecure() {
     if (keyguardManager == null) return false;
     return (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && keyguardManager.isDeviceSecure());
-  }
-
-  @VisibleForTesting
-  public boolean isDeviceSupported() {
-    return isDeviceSecure() || canAuthenticateWithBiometrics();
   }
 
   private boolean canAuthenticateWithBiometrics() {
@@ -288,18 +227,15 @@ public class LocalAuthPlugin implements MethodCallHandler, FlutterPlugin, Activi
         == BiometricManager.BIOMETRIC_SUCCESS;
   }
 
-  private void isDeviceSupported(Result result) {
-    result.success(isDeviceSupported());
+  @Override
+  public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {
+    LocalAuthApi.setup(binding.getBinaryMessenger(), this);
   }
 
   @Override
-  public void onAttachedToEngine(FlutterPluginBinding binding) {
-    channel = new MethodChannel(binding.getFlutterEngine().getDartExecutor(), CHANNEL_NAME);
-    channel.setMethodCallHandler(this);
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
+    LocalAuthApi.setup(binding.getBinaryMessenger(), null);
   }
-
-  @Override
-  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {}
 
   private void setServicesFromActivity(Activity activity) {
     if (activity == null) return;
@@ -310,11 +246,10 @@ public class LocalAuthPlugin implements MethodCallHandler, FlutterPlugin, Activi
   }
 
   @Override
-  public void onAttachedToActivity(ActivityPluginBinding binding) {
+  public void onAttachedToActivity(@NonNull ActivityPluginBinding binding) {
     binding.addActivityResultListener(resultListener);
     setServicesFromActivity(binding.getActivity());
     lifecycle = FlutterLifecycleAdapter.getActivityLifecycle(binding);
-    channel.setMethodCallHandler(this);
   }
 
   @Override
@@ -324,7 +259,7 @@ public class LocalAuthPlugin implements MethodCallHandler, FlutterPlugin, Activi
   }
 
   @Override
-  public void onReattachedToActivityForConfigChanges(ActivityPluginBinding binding) {
+  public void onReattachedToActivityForConfigChanges(@NonNull ActivityPluginBinding binding) {
     binding.addActivityResultListener(resultListener);
     setServicesFromActivity(binding.getActivity());
     lifecycle = FlutterLifecycleAdapter.getActivityLifecycle(binding);
@@ -333,7 +268,6 @@ public class LocalAuthPlugin implements MethodCallHandler, FlutterPlugin, Activi
   @Override
   public void onDetachedFromActivity() {
     lifecycle = null;
-    channel.setMethodCallHandler(null);
     activity = null;
   }