+++ packages/flutter_adaptive_scaffold/CHANGELOG.md
@@ -1,6 +1,6 @@
-## 0.2.0
+## 0.1.3
 
-* Adds `transitionDuration` parameter for specifying how long the animation should be.
+* Fixes `groupAlignment` property not available in `standardNavigationRail` - [flutter/flutter#121994](https://github.com/flutter/flutter/issues/121994)
 
 ## 0.1.2